Move ReactDataGrid component to root of src directory
