+++ src/PropTypeShapes/ExcelColumn.js
