+++ src/addons/index.js
@@ -1,4 +1,3 @@
-module.exports = require('./grids/ReactDataGrid');
 module.exports.Editors = require('./editors');
 module.exports.Formatters = require('./formatters');
 module.exports.Toolbar = require('./toolbars/Toolbar');
